move exception to dedicated module, small fixes
